requirements.txt
Pillow
PyPDF2
bullet

